# Conflicts:
#	Signal.xcodeproj/project.pbxproj
#	Signal/src/view controllers/MessagesViewController.m
#	Signal/translations/en.lproj/Localizable.strings

+++ Signal/src/views/OWSExpirationTimerView.h
@@ -1,5 +1,6 @@
-//  Created by Michael Kirk on 9/29/16.
-//  Copyright © 2016 Open Whisper Systems. All rights reserved.
+//
+//  Copyright (c) 2017 Open Whisper Systems. All rights reserved.
+//
 
 #import <UIKit/UIKit.h>
 
@@ -7,8 +8,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface OWSExpirationTimerView : UIView
 
-- (void)startTimerWithExpiresAtSeconds:(uint64_t)expiresAtSeconds
-                initialDurationSeconds:(uint32_t)initialDurationSeconds;
+- (void)startTimerWithExpiresAtSeconds:(double)expiresAtSeconds initialDurationSeconds:(uint32_t)initialDurationSeconds;
 
 - (void)stopTimer;
